Move FlingTracker.cs to the other trackers
